+++ lib/openstudio/workflow/util/measure.rb
@@ -1,5 +1,5 @@
 # *******************************************************************************
-# OpenStudio(R), Copyright (c) 2008-2020, Alliance for Sustainable Energy, LLC.
+# OpenStudio(R), Copyright (c) 2008-2021, Alliance for Sustainable Energy, LLC.
 # All rights reserved.
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions are met:
@@ -342,6 +342,8 @@ module OpenStudio
             result = nil
             begin
               load measure_path.to_s
+              # load.c in ruby can result in changing dir to root / so preserve cwd here. happens in openstudio cli 
+              Dir.chdir measure_run_dir 
               measure_object = Object.const_get(class_name).new
             rescue => e